rpi: fix framebuffer and heap memory.

+++ kernel/arch/arm/rpi/delay.c
@@ -1,6 +1,6 @@
 #include "delay.h"
 
-inline void delay(int32_t count) {
+inline void delay(ulong count) {
     asm volatile("__delay_%=: subs %[count], %[count], #1; bne __delay_%=\n"
     : "=r"(count): [count]"0"(count) : "cc");
 }
