+++ kernel/arch/arm/rpi/delay.h
@@ -1,6 +1,8 @@
 #pragma once
 
+#include <liblox/common.h>
+
 #include <stdint.h>
 
 /* Loop <count> times in a way that the compiler won't optimize away. */
-void delay(int32_t count);
+void delay(ulong count);
